Initial Commit
Recent build has been added. This is the merging of both Yaphet and
Denny's code. The Database and Models built by Travis will soon be
added.

This includes a substantial amount of the front end, along with a log in
system, badges, tags, questions, home, and course pages. Most of them
currently contain rudimentary data for development sake.